requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-CORS==6.0.0
Werkzeug>=3.1.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0